% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{Style}
\alias{Style}
\title{Style constructor with a style of no-axis y .....}
\usage{
Style()
}
\value{
a style instance with an example
}
\description{
Style constructor with a style of no-axis y .....
}
\examples{
style = Style()
style$example()
}
