% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style2.R
\name{Style2}
\alias{Style2}
\title{Style2 constructor with a style of no-axis y .....}
\usage{
Style2()
}
\value{
a style instance with an example
}
\description{
Style2 constructor with a style of no-axis y .....
}
\examples{
style = Style()
style$example()
}
