% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleTime.R
\name{StyleTime}
\alias{StyleTime}
\title{Style yearly x-axis in the Economist style}
\usage{
StyleTime(gg)
}
\arguments{
\item{gg}{a ggplot object}
}
\value{
a style instance
}
\description{
major ticks in every five years. data range is ticked.
}
\examples{
\dontrun{
gg |> StyleTime() -> style
gg + style$scale_x_continuous()
}
}
